dashboard/requirements.txt
# NTSB Aviation Accident Dashboard Requirements
# Install with: pip install -r requirements.txt

# Core Dashboard
streamlit==1.51.0

# Visualization
plotly==5.24.1
folium==0.18.0
streamlit-folium==0.23.1

# Data Processing (should already be in main .venv)
pandas>=2.0.0
numpy>=1.24.0

# Database (should already be in main .venv)
psycopg2-binary>=2.9.11

# Additional
python-dateutil==2.8.2
